pyproject.toml
[tool.poetry]
name = "blog"
version = "2.0.0"
authors = ["PauLark inc."]
description = "Blogging platform"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "^3.5.0"
asttokens = "^2.0.5"
backcall = "^0.2.0"
black = "^22.1.0"
certifi = "^2021.10.8"
charset-normalizer = "^2.0.12"
click = "^8.0.3"
coreapi = "^2.3.3"
coreschema = "^0.0.4"
decorator = "^5.1.1"
Deprecated = "^1.2.13"
Django = "^4.0.2"
django-environ = "^0.8.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.0.0"
drf-yasg = "^1.20.0"
environ = "^1.0"
executing = "^0.8.2"
idna = "^3.3"
inflection = "^0.5.1"
ipython = "^8.0.1"
itypes = "^1.2.0"
jedi = "^0.18.1"
Jinja2 = "^3.0.3"
MarkupSafe = "^2.1.0"
matplotlib-inline = "^0.1.3"
mypy-extensions = "^0.4.3"
packaging = "22.0"
parso = "^0.8.3"
pathspec = "^0.9.0"
pexpect = "^4.8.0"
pickleshare = "^0.7.5"
setuptools = "^72.1.0"
pillow = "^10.4.0"
platformdirs = "3.0.0"
prompt-toolkit = "^3.0.28"
psycopg2 = "^2.9.3"
ptyprocess = "^0.7.0"
pure-eval = "^0.2.2"
Pygments = "^2.11.2"
PyJWT = "^2.3.0"
pyparsing = "^3.0.7"
pytz = "^2021.3"
redis = "^4.1.4"
requests = "^2.27.1"
ruamel-yaml = "^0.17.21"
ruamel-yaml-clib = "^0.2.6"
six = "^1.16.0"
sqlparse = "^0.4.2"
stack-data = "^0.2.0"
tomli = "^2.0.1"
traitlets = "^5.1.1"
typing_extensions = "^4.1.1"
uritemplate = "^4.1.1"
urllib3 = "^1.26.8"
wcwidth = "^0.2.5"
wrapt = "^1.13.3"
